refactor: share versioned card schemas
