+++ docs/agents/architecture.md
@@ -33,7 +33,7 @@ Before changing background commands or sync execution, read [ADR-0001](../adr/00
 - Cards are slug-keyed, own their notes, and store dates as numbers.
 - Preserve supported card fields, FSRS schedules, notes, stats, settings, and Gist configuration. Unrecognized fields may be discarded; unrelated undecodable records need not survive mutations.
 - Schema changes require a new numbered module in `infrastructure/storage/migrations/`, appended to the sequence in `infrastructure/storage/migrations/runner.ts`.
-- Each migration declares its historical input and output contracts. A loader validates the preceding migration's output shape; its saver accepts its own output type. Keep historical validation and storage keys independent of current application models and layout changes.
+- Each migration declares its historical input and output contracts. A loader validates the preceding migration's output shape; its saver accepts its own output type. Shared schemas live in `domain/schemas/vN.ts`, named for the schema version that introduced them. Keep existing contracts fixed; reuse or extend them for later versions. Application models and migrations compose these definitions with their own normalization and extra-field handling. Migrations import a specific version, never current application models; historical storage keys remain independent of layout changes.
 - Use `Input`, `Output`, `validateInput`, and `validateOutput` consistently within each migration. Export `Output` and `validateOutput` for the next migration, which imports them as `Input` and `validateInput`. When a local input validator adds checks or error context, import the preceding validator as `validatePreviousOutput` and call it from `validateInput`. For an identity migration, use `export type Output = Input` and alias the validator with `export const validateOutput: typeof validateInput = validateInput`. Validators accept `unknown` and assert their validated type; local helper types may express additional narrowing. Keep the lifecycle names `load`, `migrate`, `save`, and optional `cleanup`.
 - A physical layout reader knows where values are stored and how to assemble them into a raw dataset; it does not validate a migration's data schema. Each migration's loader validates those raw values against its historical input contract. Introduce a new layout reader when storage locations or assembly rules change.
 - Startup validates the completed version, then loads, transforms, saves, and cleans up each step before recording its version. The runner assumes no dataset shape. Backup parsing validates the declared version's contract before running the remaining pure transformations, including when none remain. Missing version means version 0.
