+++ domain/notes.ts
@@ -1,10 +1,9 @@
 import { z } from 'zod';
+import { noteTextSchema } from './schemas/v4';
 
-export const NOTES_MAX_LENGTH = 500;
+export { NOTES_MAX_LENGTH } from './schemas/v4';
 
 export const noteSchema = z.object({
-  text: z.string().max(NOTES_MAX_LENGTH, {
-    error: `Note exceeds maximum length of ${NOTES_MAX_LENGTH} characters`,
-  }),
+  text: noteTextSchema,
 });
 export type Note = z.infer<typeof noteSchema>;
